ai-backend/pyproject.toml
[project]
name = "ai-backend"
version = "0.1.0"
description = "AI backend service for DuckTodo using FastAPI and LangChain"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pymysql>=1.1.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "PyJWT>=2.8.0",
]
